ABINet/requirements.txt
torch==1.1.0
torchvision==0.3.0
fastai==1.0.60
LMDB
Pillow<7.0.0
opencv-python
tensorboardX